relocate run methods from container to image
