+++ README.md
@@ -58,13 +58,13 @@ More info [here](http://conu.readthedocs.io/en/latest/installation.html).
 ```python
 # TODO: use more complex example with assertions and stuff
 from conu.backend.docker.image import DockerImage
-from conu.backend.docker.container import DockerContainer, DockerRunBuilder
+from conu.backend.docker.container import DockerRunBuilder
 
 # initialization
 image = DockerImage('fedora', tag='26')
 image.pull()
 command = DockerRunBuilder(command=["ls"], additional_opts=["-i", "-t"])
-container = DockerContainer.run_via_binary(image, command)
+container = image.run_via_binary(command)
 
 # cleanup
 container.stop()
