+++ conu/apidefs/image.py
@@ -118,6 +118,33 @@ class Image(object):
         """
         raise NotImplementedError("mount is not implemented")
 
+    def run_via_binary(self, *args, **kwargs):
+        """
+        create container using image and run it in the background; this method is useful
+        to test real user scenarios when users invoke containers using binary and not an API
+
+        :param image: instance of Image
+        :return: instance of Container
+        """
+        raise NotImplementedError("run_via_binary method is not implemented")
+
+    def run_via_api(self, container_params):
+        """
+        create container using image and run it in the background
+
+        :param container_params: instance of ContainerParameters
+        :return: instance of Container
+        """
+        raise NotImplementedError("run_via_api method is not implemented")
+
+    def create(self, container_params):
+        """
+        create container using image
+
+        :param container_params: instance of ContainerParameters
+        :return: instance of Container
+        """
+        raise NotImplementedError("create method is not implemented")
 
 class S2Image:
     """
