+++ docs/source/examples/s2i.py
@@ -1,10 +1,10 @@
-from conu import S2IDockerImage, DockerContainer
+from conu import S2IDockerImage
 
 source = 'https://github.com/dbarnett/python-helloworld'
 image = S2IDockerImage("centos/python-35-centos7")
 image.pull()
 extended_image = image.extend(source, "myapp")
-container = DockerContainer.run_via_binary(extended_image)
+container = image.run_via_binary()
 
 container.stop()
 container.delete()
