+++ docs/source/reference/docker_container.rst
@@ -4,7 +4,7 @@ Docker Container
 Aside from methods in API definition - :class:`conu.apidefs.container.Container`, DockerContainer implements following methods:
 
 .. autoclass:: conu.DockerContainer
-   :members:  inspect, wait_for_port, run_via_binary_in_foreground
+   :members:  inspect, wait_for_port
 
 .. autoclass:: conu.DockerRunBuilder
    :members:
